Adding Blind RRT, Radial Blind RRT, environments that I used for testing and xmls
git-svn-id: svn+ssh://parasol-svn.cs.tamu.edu/research/parasol-svn/svnrepository/pmpl/branches/blindRRT@3923 5e2c2758-810a-0410-8eb3-8e4e92d4004e

+++ src/Cfg/Cfg.cpp
@@ -99,7 +99,8 @@ Cfg::operator==(const Cfg& _cfg) const {
     return false;
   for(size_t i = 0; i < m_dof; ++i) {
     if(m_dofTypes[i] == POS || m_dofTypes[i] == JOINT) {
-      if(fabs(m_v[i]-_cfg[i]) > numeric_limits<double>::epsilon())
+      // TODO Cesar if(fabs(m_v[i]-_cfg[i]) > numeric_limits<double>::epsilon())
+      if(fabs(m_v[i]-_cfg[i]) > numeric_limits<float>::epsilon())
         return false;
     }
     else {
